02_bioos_engine/simulations/demo_scripts/requirements.txt
biopython
qutip
numpy
matplotlib
plotly
dna_features_viewer
tqdm